requirements.txt
pandas
langchain
langchain-community
langchain-google-genai
unstructured
python-dotenv
streamlit
llama-index-core
llama-parse
markdown
nest-asyncio
opensearchpy